requirements.txt
google-cloud-vision
certifi
Pillow
mss
numpy

